+++ clients/admin-ui/src/features/dataset/EditDrawer.tsx
@@ -75,7 +75,7 @@ const EditDrawer = ({
       <ConfirmationModal
         isOpen={deleteIsOpen}
         onClose={onDeleteClose}
-        onDelete={onDelete}
+        onConfirm={onDelete}
         title={deleteTitle}
         message={deleteMessage}
       />